Update Budget.js
In the above code snippet we are using the useState hook to create a state variable called newBudget and initialize it with the current value of budget. We are also defining a function called handleBudgetChange that updates the value of newBudget when the user changes the value of the input field.

We are then setting the value attribute of the input field to newBudget and adding an onChange event listener that calls handleBudgetChange when the user changes the value of the input field.

Here, I'm using the Bootstrap Alert classes to give a nice gray background by adding some text and hard coding a value.